pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supervisory-procedures"
version = "0.1.0"
description = "Skill-as-Code platform for AI agent governance"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["ai", "governance", "agents", "compliance"]

dependencies = [
    "click>=8.1",
    "questionary>=2.0",
    "pyyaml>=6.0",
    "jsonschema[format-nongpl]>=4.21",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.0",
    "types-PyYAML",
    "types-jsonschema",
]

[project.scripts]
supv = "supervisory_procedures.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["supervisory_procedures*"]

[tool.setuptools.package-data]
supervisory_procedures = ["py.typed", "../schema/*.json"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
